# KaniniAcademy Backend/KaniniAcademy/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "abhi",
    "version": "v1"
  },
  "paths": {
...
    "/api/Certificate/{id}": {
      "delete": {
        "tags": [
          "Certificate"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Certificate"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Certificate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Certificate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Certificate"
              }
            }
          }
        }
      }
    },
...
    "/api/CtcReg": {
      "get": {
        "tags": [
          "CtcReg"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CtcReg"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CtcReg"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CtcReg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CtcReg/{id}": {
      "delete": {
        "tags": [
          "CtcReg"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CtcReg"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CtcReg"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CtcReg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CtcReg"
              }
            }
          }
        }
      }
    },
    "/api/FreeResource": {
      "get": {
        "tags": [
          "FreeResource"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FreeResource"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FreeResource"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FreeResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FreeResource/{id}": {
      "delete": {
        "tags": [
          "FreeResource"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FreeResource"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FreeResource"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FreeResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FreeResource"
              }
            }
          }
        }
      }
    },
...
    "/api/Points": {
      "get": {
        "tags": [
          "Points"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Point"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Points"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Point"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Points/{id}": {
      "delete": {
        "tags": [
          "Points"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Point"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Points"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Point"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Point"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Certificate": {
      "type": "object",
      "properties": {
        "cid": {
          "format": "int32",
          "type": "integer"
        },
        "cname": {
          "format": "int32",
          "type": "integer"
        },
        "duration": {
          "type": "string"
        },
        "projects": {
          "type": "string"
        },
        "videos": {
          "format": "byte",
          "type": "string"
        },
        "image": {
          "format": "byte",
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "keySkills": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "howItWorks": {
          "type": "string"
        },
        "syllabus": {
          "type": "string"
        },
        "faqQues": {
          "type": "string"
        },
        "faqAns": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CtcReg": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "tid": {
          "format": "int32",
          "type": "integer"
        },
        "cid": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "FreeResource": {
      "type": "object",
      "properties": {
        "freeRid": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "providedBy": {
          "format": "int32",
          "type": "integer"
        },
        "downloads": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "thumbnail": {
          "format": "byte",
          "type": "string"
        },
        "book": {
          "type": "string"
        },
        "template": {
          "type": "string"
        },
        "video": {
          "format": "byte",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Point": {
      "type": "object",
      "properties": {
        "uid": {
          "format": "int32",
          "type": "integer"
        },
        "qid": {
          "format": "int32",
          "type": "integer"
        },
        "points": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}